requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.26.0
Django==3.2.5
django-cloudinary-storage==0.3.0
djangorestframework==3.12.4
idna==2.10
Pillow==8.3.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
